+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add tests for `VersionConfig.{filter_targets,filter_dates}`
+- Add tests for `VersionConfig.{filter_targets,filter_dates,filter_people}`
 
 ### Fixes